backend/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
filelock==3.18.0
flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
flask-sqlalchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
importlib-metadata==8.6.1
importlib-resources==6.5.2
itsdangerous==2.2.0
jinja2==3.1.6
keras==3.9.2
kiwisolver==1.4.7
libclang==18.1.1
markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
opt-einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==11.2.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
scipy==1.13.1
six==1.17.0
sqlalchemy==2.0.40
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
torch==2.7.1
torchvision==0.22.1
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
